Refactored pdf module into a sub-package
