+++ docs/source/images/gallery_images/gallery_density_estimation.py
@@ -1,6 +1,6 @@
 from numpy import linspace
 from numpy.random import normal, random, exponential
-from inference.pdf import UnimodalPdf
+from inference.pdf.unimodal import UnimodalPdf
 import matplotlib.pyplot as plt
 
 N = 5000
