+++ inference/pdf/__init__.py
@@ -1,0 +1,13 @@
+from inference.pdf.base import DensityEstimator
+from inference.pdf.kde import GaussianKDE, KDE2D
+from inference.pdf.unimodal import UnimodalPdf
+from inference.pdf.hdi import sample_hdi
+
+__all__ = [
+    "DensityEstimator",
+    "GaussianKDE",
+    "KDE2D",
+    "UnimodalPdf",
+    "sample_hdi"
+]
+
